requirements.txt
yfinance
pandas
matplotlib
jinja2
beautifulsoup4
curl-cffi
pyyaml
openai
